+++ ETMS.Entity/Database/Source/EtStudentCourse.cs
@@ -49,7 +49,7 @@ namespace ETMS.Entity.Database.Source
         /// <summary>
         /// 消耗数量
         /// </summary>
-        public int UseQuantity { get; set; }
+        public decimal UseQuantity { get; set; }
 
         /// <summary>
         /// 消耗单位  <see cref="ETMS.Entity.Enum.EmCourseUnit"/>
@@ -59,12 +59,12 @@ namespace ETMS.Entity.Database.Source
         /// <summary>
         /// 剩余数量(课时/月)
         /// </summary>
-        public int SurplusQuantity { get; set; }
+        public decimal SurplusQuantity { get; set; }
 
         /// <summary>
         /// 剩余数量(天)
         /// </summary>
-        public int SurplusSmallQuantity { get; set; }
+        public decimal SurplusSmallQuantity { get; set; }
 
         /// <summary>
         /// 课程分析JOB最后执行时间
@@ -89,6 +89,6 @@ namespace ETMS.Entity.Database.Source
         /// <summary>
         /// 超上课时
         /// </summary>
-        public int ExceedTotalClassTimes { get; set; }
+        public decimal ExceedTotalClassTimes { get; set; }
     }
 }
