+++ ETMS.Entity/View/ClassRecordAbsenceLogView.cs
@@ -109,7 +109,7 @@ namespace ETMS.Entity.View
         /// <summary>
         /// 扣的课时
         /// </summary>
-        public int DeClassTimes { get; set; }
+        public decimal DeClassTimes { get; set; }
 
         /// <summary>
         /// 课消金额
@@ -119,7 +119,7 @@ namespace ETMS.Entity.View
         /// <summary>
         /// 超上课时
         /// </summary>
-        public int ExceedClassTimes { get; set; }
+        public decimal ExceedClassTimes { get; set; }
 
         /// <summary>
         /// 点名时间
