+++ ETMS.Entity/View/StudentCourseView.cs
@@ -59,7 +59,7 @@ namespace ETMS.Entity.View
         /// <summary>
         /// 消耗数量
         /// </summary>
-        public int UseQuantity { get; set; }
+        public decimal UseQuantity { get; set; }
 
         /// <summary>
         /// 消耗单位  <see cref="ETMS.Entity.Enum.EmCourseUnit"/>
@@ -69,12 +69,12 @@ namespace ETMS.Entity.View
         /// <summary>
         /// 剩余数量(课时/月)
         /// </summary>
-        public int SurplusQuantity { get; set; }
+        public decimal SurplusQuantity { get; set; }
 
         /// <summary>
         /// 剩余数量(天)
         /// </summary>
-        public int SurplusSmallQuantity { get; set; }
+        public decimal SurplusSmallQuantity { get; set; }
 
         /// <summary>
         /// 课程分析JOB最后执行时间
